Added basic scripting foundation (disabled)
added basic scripting foundation code. All of the scripting code is
wrapped in #ifdef SMASHBROS_SCRIPT_DISABLE and the script is by
default, disabled. ScriptManager and the modules have not yet been
added to the code from the windows version

+++ Source/SmashBros/Game/Stages/BattlefieldBrawl.cpp
@@ -3,7 +3,7 @@
 
 namespace SmashBros
 {
-	BattlefieldBrawl::BattlefieldBrawl(int x1, int y1) : Stage(x1, y1)
+	BattlefieldBrawl::BattlefieldBrawl(float x1, float y1) : Stage(x1, y1)
 	{
 		setSpawnPoint(0, 0,0);
 		setSpawnPoint(1,  115, -70);
