+++ Source/SmashBros/Game/Stages/HyruleTemple.h
@@ -14,11 +14,11 @@ namespace SmashBros
 		void loadForeground();
 		void loadHangPoints();
 		void loadGround();
-
+		
 	public:
-		HyruleTemple(int x1, int y1);
+		HyruleTemple(float x1, float y1);
 		~HyruleTemple();
-
+		
 		virtual void Update(long gameTime);
 		virtual void Draw(Graphics2D&g, long gameTime);
 	};
